copy attachments with their original file names
See: https://www.phpbb.com/support/docs/en/3.0/kb/article/backing-up-attachments-with-their-original-filenames/
Both the filename used by phpBB and the original filename are stored in the "attachments" table, so to recover the original filename you can use the data from that table.

Here is a script which copies all files (except thumbnails) uploaded as attachments to a different folder and renames them to the original filename, prefixed by the attachment_id to prevent duplicate file names. The original data in phpBB is left untouched.